requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
environs==14.5.0
fastapi==0.128.0
fastapi_cors==0.0.6
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
kombu==5.6.2
marshmallow==4.2.0
packaging==25.0
playwright==1.57.0
playwright-stealth==2.0.0
prompt_toolkit==3.0.52
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.1.0
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.40.0
vine==5.1.0
wcwidth==0.2.14
psycopg[binary]==3.3.2
